test: migrate uncollected claude_cli tests into CI-collected tests/ tree
src/worca/utils/test_claude_cli.py held 43 test methods across 6 classes
but was never collected by CI (testpaths = ["tests"]). Migrated ~22 unique
cases into tests/test_claude_cli.py and a new tests/test_claude_cli_logging.py,
fixed the unmocked _resolve_tool_args reach that caused 4 failures in clean
environments, dropped ~22 redundant tests already covered by the collected
suite, deleted the source file, and added a meta-test
(tests/test_no_stray_tests.py) to prevent recurrence.
